+++ lib/widgets/two_way_binding_builder.dart
@@ -2,12 +2,22 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_essentials_kit/errors/localized_error.dart';
 import 'package:flutter_essentials_kit/misc/two_way_binding.dart';
 
-class TwoWayBindingWidget<T> extends StatelessWidget {
-  final TwoWayBinding<T> binding;
-  final Widget Function(BuildContext context, T data, Function(T) onChange,
-      LocalizedError error) builder;
+/// Typedef of [TwoWayBindingBuilder].
+/// The [data] and [error] can be null. The other parameters are not-null.
+typedef TwoWayBindingDataBuilder<T> = Widget Function(
+  BuildContext context,
+  TextEditingController controller,
+  T data,
+  Function(T) onChange,
+  LocalizedError error,
+);
 
-  TwoWayBindingWidget({
+/// Widget that builds itself based on the latest data of a [TwoWayBinding].
+class TwoWayBindingBuilder<T> extends StatelessWidget {
+  final TwoWayBinding<T> binding;
+  final TwoWayBindingDataBuilder<T> builder;
+
+  TwoWayBindingBuilder({
     Key key,
     @required this.binding,
     @required this.builder,
@@ -21,6 +31,7 @@ class TwoWayBindingWidget<T> extends StatelessWidget {
         stream: binding.stream,
         builder: (context, snapshot) => builder(
           context,
+          binding.textController,
           snapshot.data,
           binding.change,
           snapshot.error,