+++ test/misc/two_way_binding_test.dart
@@ -15,7 +15,6 @@ void main() {
 
   test('change with some values', () async {
     binding = binding.onChange((value) {
-      print(value);
       expect(values.contains(value), isTrue);
     });
 
@@ -24,11 +23,46 @@ void main() {
     await futureExpect;
   });
 
-  test('onChange with some values', () {
+  test('value with some values', () async {
     binding = binding.onChange((value) {
-      print(value);
-
-
+      expect(values.contains(value), isTrue);
     });
+
+    final futureExpect = expectLater(binding.stream, emitsInOrder(values));
+    values.forEach((value) {
+      binding.value = value;
+      expect(binding.value, equals(value));
+    });
+    await futureExpect;
+  });
+
+  test('bindDataRule with success', () async {
+    binding = binding.bindDataRule(RequiredRule());
+
+    final futureExpect = expectLater(binding.stream, emitsInOrder(values));
+    values.forEach(binding.change);
+    await futureExpect;
+  });
+
+  test('bindDataRule with error', () async {
+    binding = binding.bindDataRule(RequiredRule());
+
+    final futureExpect = expectLater(
+        binding.stream, emitsError(isInstanceOf<RequiredRuleError>()));
+    binding.change(null);
+    await futureExpect;
+  });
+
+  test('bindDataRule2 with error', () async {
+    final binding2 = TwoWayBinding<String>().bindDataRule2(binding, SameRule());
+
+    final futureExpect = expectLater(
+      binding2.stream,
+      emitsError(isInstanceOf<SameRuleError>()),
+    );
+
+    binding.value = null;
+    binding2.value = '';
+    await futureExpect;
   });
 }